requirements.txt
flask
flask_login
flask_sqlalchemy
jinja2
matplotlib
sqlalchemy
flask_cors
flask_caching
pandas
flask_restful
redis
celery
PyJWT